Refactor and update acceptance tests around saving settings.
