+++ cms/djangoapps/contentstore/features/grading.feature
@@ -32,7 +32,7 @@ Feature: Course Grading
         And I have populated the course
         And I am viewing the grading settings
         When I change assignment type "Homework" to "New Type"
-        And I save my changes
+        And I press the "Save" notification button
         And I go back to the main course page
         Then I do see the assignment name "New Type"
         And I do not see the assignment name "Homework"
@@ -42,6 +42,7 @@ Feature: Course Grading
         And I have populated the course
         And I am viewing the grading settings
         When I delete the assignment type "Homework"
+        And I press the "Save" notification button
         And I go back to the main course page
         Then I do not see the assignment name "Homework"
 
@@ -50,7 +51,7 @@ Feature: Course Grading
         And I have populated the course
         And I am viewing the grading settings
         When I add a new assignment type "New Type"
-        And I save my changes
+        And I press the "Save" notification button
         And I go back to the main course page
         Then I do see the assignment name "New Type"
 
@@ -59,5 +60,22 @@ Feature: Course Grading
         And I have populated the course
         And I am viewing the grading settings
         When I change assignment type "Homework" to "New Type"
-        And I cancel my changes
+        And I press the "Cancel" notification button
         Then I do not see the changes persisted on refresh
+
+    Scenario: Confirmation is shown on save
+        Given I have opened a new course in Studio
+        And I have populated the course
+        And I am viewing the grading settings
+        When I change the "<field>" field to "<value>"
+        And I press the "Save" notification button
+        Then I see a confirmation that my changes have been saved
+        # Lettuce hooks don't get called between each example, so we need
+        # to run the before.each_scenario hook manually to avoid database
+        # errors.
+        And I reset the database
+
+    Examples:
+        | field                          | value               |
+        | Course Grading Graceperiod     | 1:00                |
+        | Course Grading Assignment Name | New Assignment Name |
